sharing [nfc]: Deduplicate code b/w Pm and Stream sharing screens.
Deduplication done as much as possible, by creating a `ShareWrapper`
component which wraps around the components differing in `ShareToPm`
and `ShareToStream`.

`ShareWrapper` contains the common stylesheet, components and the
logic to send/cancel shared data.

To do this:

* The common components were identified in render method of
both `ShareToPm` and `ShareToStream`, they were moved into the file
`src/sharing/ShareWrapper.js`.
* Associated data was also moved and removed from both the files.
* `handleSend` method was modified to contain the code of
  `handleSend` function present in `src/sharing/send.js`.
* This file was present only for code deduplication purposes and
  hence was no longer relevant, therefore it was deleted.
